added fuzzy search extension
